"drop-hostkeys-message" for sshpiper
this is an obscure flag to sshpiper that avoids
"client_global_hostkeys_prove_confirm: server gave bad signature for
ED25519 key 0: incorrect signature" warnings on the client.

ChatGPT explained it to me as:

	That message comes from OpenSSH’s UpdateHostKeys path, after the normal key exchange succeeds.

You can sort of see it with -vvv. Looks like SSH has some features for
letting the server rotate a host key, and passing this through to the
upstream doesn't make sense because we have different host keys.
Spelunking in the sshpiper source code found that this already had
a workaround.