refactor w/ new billing pkg, mock stripe
This change adds a new billing package, which contains
a Billing interface and functions to construct instances of
it with default or mock stripe service dependencies.

See billing/README.md for instructions on how to use the
NewWithMockStripe function in unit tests, and how to run a separate
standalone mock stripe server and point your dev exed instance
at it using environment variables.