container: autostart exe sshd where we can
For exeuntu, instead of showing up later with a `nerdctl exec`
of sshd, we can set the entrypoint directly to /exe.dev/bin/sshd.
This, on a happy start, cuts `ssh localexe new` start time from
3s to 2s! (My best time so far is 1.8s.)

There are probably more containers we can safely do this for,
but I'm starting with exeuntu because we understand it.