tini: modify to start /exe.dev/bin/sshd
Besides starting exe.dev's sshd and then operating as a typical
init, we add one feature: if the followon program looks like some
other init, in particular one that will not take well to being
wrapped, we exec it right over us without a fork, effectively
passing control to the other init.

In this way, we can make this the ENTRYPOINT of *any* container,
just by prepending "exetini", "-g", "--" to their existing
entrypoint (or to "sleep infinity") if they have no entrypoint+cmd.