all: use package sqlite
We had package sqlite available...but none of the code used it.

As a consequence, under concurrent load (which conveniently e2e/expect
created, even in its nascent state), we got "database is busy" failures.

This is a giant, sprawling, semi-vibe-coded migration to actually
use package sqlite, which also required threading some contexts
through and using errors.Is to check for sql.ErrNoRows.

The work is unfinished, but I want to get it in before it rots.
(It took hours, between me and the LLMs.)

Remaining items to do:

- Finish auditing all of the contexts in use; many have the wrong
  lifetime and/or should be threaded through. I started on this
  but ran out of steam, as it is extraordinarily tedious work.
- Remove all instances in the tests of directly opening databases;
  this will end badly, and we may as well stop doing it now.
- Add pre-commit checks to enforce the above work to avoid regressions.