container: CI fixes
Some of this is IP ranges for container tests
(no exed to determina a range for it).
Some of it is timeout stuff (ugh).
We also adjust the clone of the /exe.dev drive for new containers so
if running on an FS that doesn't support cloning (like our CI machine),
fall back to traditional copy.