container, ops: use kata snapshot/restore to start VMs
This, in theory, saves time on startup because some of the VM
is already baked. The tradeoff is some I/O copying the memory
of the snapshotted VM image, and colima performance is noisy
enough that it's hard to be sure it is actually faster. But I
believe it is getting us to a good state in a more consistent
amount of time.

(It will also be interesting to test this on real metal
hardware.)