shelley: first steps
Shelley (after the poet and, um, well, the unix shell) is the little
agent that could.

I believe, currently, that to get a good exe.dev experience, we'll
need to customize the agent, and, to do so, we'll need to have
an agent. So I've used some learnings from Sketch to build a little
one.

The highlights are:
 - There's a new "predictable" model that you can use for testing,
   that lets you script some interactions without bothering to do RPCs.
   (This is just a text fixture.)
 - There's both a CLI and an HTTP server, but they can share the same
   database. This is mostly useful for testing right now.
 - When it's not the agent's turn, the conversation is on disk (sqlite).
   This means the agent can be stopped and restarted and continue
   where it left of.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: s7696d5703c79de20k