container: directly inspect image in remote
We were trying to ask nerdctl on the host to give us the image
manifest. This is surprisingly difficult! Nydus has all the layers
to start a container, but the manifest comes and goes.

Instead, we go straight to the registry and request the manifest.
This is reliable, and let's us cut out a total hack that got
added by an LLM.

Also pull some image size nonsense off the critical path.