container: add support for local sha256: image refs
if you run 	"new --image sha256:..." with the sha of an image that you have
locally built and then loaded into lima-exe-ctr's containerd, it should
create a new box using that local image.

The Create logic in NerdctlManager now bypasses the tagManager when the
container ref starts with "sha256:" and looks for it locally instead of
asking remote container repos for it.