exeuntu: add systemd
Maintaining multiple container images seems too painful; let's try
just having one. The downside is that you need more flags to just
run the image, but that's fine.

The fact that the entrypoint script is called "initd" is because
that's what "tini" expects when it makes a decision about whether
to fork or exec. That's not a perfect design at the moment, but we'll
deal with it.