ci: grep -q trivia
I want ci-vm-start to work for me on the CI machine, but "grep -q" and
set -o pipefail ends up exiting grep before the input is done, which
ends up with an error, and goes down the wrong branch. No idea why
I see this interactively but not in CI, but fixing regardless.