exedb schema changes for llmgateway accounting
this set of changes is big, but it shouldn't alter the functionality of
exed much, if at all, to an end user.

this lays the groundwork for another set of changes for llmgateway that
make it actually charge the billing account for a box's alloc whenever it
proxies a response that contains billable details from the origin llm service.

I had to update some unrelated tests so that they use the new
Server.createUser method instead of making direct calls to db.Tx and Rx, since
those were not creating a complete, correct user account in the db.