ghuser: validate ssh keys in band
So if a user revokes a leaked key, we pick that up.

Adds a round trip, which is a real bummer, but important.

We can get one round trip back in the future
by actively maintaining the db to include
a GitHub login and other useful data.

Not hard, just have to build it if/when we decide it matters.