exe: add logged-out page
To avoid a confusing apparently infinite loop when you click "log out".

It used to redirect you to back to /, which immediately
prompted you to log back in. Pretty confusing to click "log out"
and hit a page asking you to log in.

Not entirely sure this is the best long term solution,
but it's simple and better than what we have now.