web: add web shell at /shell with xterm.js and websockets
There's some code duplication with the xterm stuff on the frontend,
but that's harmless enough for now.

The idea here is kinda cute; we already have a terminal interface,
let's duplicate it on the web.

The main "issue" was figuring out the subset of ssh.Session that we
need, and implementing that on both sides. Nothing that the agent
couldn't handle.

Note that I'm using websockets here rather than SSE-with-a-separate-POST
endpoint, because I think that's ultimately cleaner here, but who knows.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: s112e050b4f350fd8k