web: consolidate on /~ for logged in list of boxes
We have .plan, finger, and so many more options!

Anyway, I've combined the two existing list-of-boxes pages
into one, and made links to Shelley as well as to the other
things we've got going.

Mobile is vaguely ok. I made a conscious decision to hide the
SSH link on mobile, since what are you going to do with that?

I made a user page to stick the e-mail address and keys on.
We'll need it later for adding passkeys, changing e-mail addresses,
etc.

The top bar navigation for logged in users needs more love, and also
needs love on mobile. I didn't tackle it.