shelley<->exed: add --prompt option to new command for shelley integration
The idea is that "new --prompt=whatever" works. We do this by talking to
Shelley over HTTP and then interpreting the stream. This ties us
version-wise with the "current" exeuntu release, but that's fine. We
don't have to deal with old exeuntu's because you can't run "new" on
them.

The misery was of course getting Claude to figure out what end of turn
meant, and I had to change Shelley to handle that, and the test is
currently skipped because of the aforementioned versioning stuff.