container: add another fast path to verifyKataRuntime
Courtesy of codex. I don't understand any of this.
But on my mac, this cuts 5s off of e1e tests.

I kinda suspect that expecting 'nerdctl info --format json'
to contain the string 'kata' may be wrong and the check should be deleted.
(It doesn't appear to list plug-ins at all.)
But again: I don't understand any of this. So I left it alone.