.github: fix handling of broken rebases
The mid-flight 'git rebase -abort' returns git to the commit
it was at when the rebase was attempted...which is definitely
NOT necessarily the commit we're trying to push!

Do the rebase fix-up much earlier, before doing the checkout.
Sanity-check it later.
While we're here, fix the argument order to merge-base.