cobble: prevent nil pointer dereference in laters.run
On the first call to laters.run, the laters slice was cleared, setting
all funcs to nil, and so on subsequent calls to laters.run, it would
attempt to call nil funcs, causing a panic. This change sets the laters
slice to nil after clearing it, preventing further calls to run from
attempting to call nil funcs.