exeuntu: move motd from /etc/update-motd.d to ~/.bashrc
ubuntu uses PAM and /etc/update-motd.d for dynamic MOTD. Our SSH isn't
compiled with PAM, and also we can't depend on that in general,
so let's resort to the lowest common denominator, which is something
in ~/.bashrc. On the plus side, it's probably easier for a user to turn
off when they get annoyed.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: se46ded2b19ad80d5k