e1e: use busybox instead of python3 to serve http
There's a comment in the code saying it doesn't work...
but it does work. And at least on my machine, cuts the runtime
of these tests from 70-80s to 40-50s.

Those numbers include some other fiddling I'm doing which isn't in this branch.
Key point is just: It's way faster.