e1e: unify TestSSHWorks and TestDockerWorks
Now they are subtests of TestVanillaBox.

The guiding principle here is that we should re-use boxes
in tests when it is reasonable to do so, for efficiency.

It's not hard to arrange for different tests of a box to use different
parts of the box itself. The challenge is around box config itself.
These tests both use a vanilla, unmodified box; share it.