e1e: unify proxy tests
This allows them to re-use a box, which should improve efficiency.
It allows some code re-use.

Along the way, we fix a bug:
There was an attempt to wait for the http server to be running.
However, it started the command and waited for it to finish...
without checking whether it succeeded. It did not, as it was malformed.
Fix both those bugs.

It seems (naively, to me) like we ought to be able to re-use even more
code, as the only real difference is the port in question.
Leave that as a TODO.