llmgateway: strip .exe.dev from hostname.
Shelley is busted currently because of this. The hostname
story is weird.

I want the box to know that "hostname -f" is "box.exe.dev", because
that lets the user have good URLs and stuff. The "traditional" way
of doing this seems to be to set /etc/hostname to "box" and then have
an entry in "/etc/hosts" that maps box.exe.dev to the same IP that box
has, whatever that is. Instead of that, I earlier wrote "box.exe.dev" to
/etc/hostname (by passing --hostname to nerdctl), and /etc/hosts is
its own shambles (that nerdctl makes).

Anyway--for now, let's just strip it. I'm not sure what the right right
answer is here.