execore: accept an A record as proof for apex domains
This allows a user to set an ALIAS record for their apex
domain (e.g. knownhosts.net) to point to their box
(e.g. knownhosts.exe.dev). This is translated into a DNS
A record pointing at one of exed's public IPs. If we see
this, accept it as evidence of ownership and fetch a
Let's Encrypt certificate for the apex domain.