ops: make setup-lima-hosts.sh quieter
This is going to go away soon anyway,
but I had to wade through this a lot while
working on 04dde8571c5f22ebc67b22d8118297c240b02461,
so I had plenty of time to trim it down anyway.

It's still noisier than I would like, but it's a start.