execore: drive repl `ls` off db, not ctrhost
These should be in sync.
We don't enforce that, but we will (soon?).

There is a slight loss of fidelity here:
We store in the image name in the db without the sha256
of the image hash, which is what causes the golden test changes.
This is probably not critical; if we decide it's important,
we can start storing the sha256 in the db.

This does mean that we lose some test coverage of listing containers,
which is a pity for the exelet work.
The right way to get that back is probably to implement the
state comparison logic, and run it at the end of the e1e tests.