execore: make "failed container" message less fail-y
And fix the recommended command.

The deeper need here is to think about the various failure states:

- didn't start
- shut down by user
- moved to trash but restorable
- permanently destroyed

And then distinguish them and provide better,
more distinctive error messages for each.

For now, making the text generic will at least make it not wrong.

(This is in response to feedback in Discord from jmandel.)