xterm: add persistent sessions via dtach with color names
Ok. Now we do the fun thing and use "dtach". We name sessions
after colors and they're in the URL bar. So, you can copy the
URL and get the same terminal. Or you can open new windows, and
it'll create a new terminal.

This also fixes two subtle bugs:

1. The terminal focuses by default.

2. The terminal redraws.

Redrawing was hard. But if you send a terminal resize command,
and it's not the same size as you already have, it seems to work.
So we send a resize to (h+1, w+1) and then a resize to (h, w),
and it works. Sometimes. If you wait in between them, it seems
to work most of the time. So we wait in between, but do it cleverly
so mostly we don't wait much.t

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: se245d86eb867769ek