route53: stage all challenges before processing them
The challenges for exe.dev and *.exe.dev share a DNS entry.

If we process them one at a time, we churn that DNS entry,
which (depending on TTLs and propagation and luck)
can cause LetsEncrypt to fail.

Instead, stage all the challenges,
then process them, then tear them down.