exelet: curl http://169.254.169.254/
I want the boxes to have some way of communicating to the exelet.
Ultimately, I'm very excited to have the LLM gateway going through
box -> exelet -> exe.dev without any auth tokens, because we can
trust the remote IP on that first hop. But, first things first,
this establishes an HTTP server on the exelet that's listening
on 169.254.169.254:80.

Needless to say, there's no configuration here to make it possible
to run multiple exelet's on one host. Ultimately, we'd need to
parameterize the bridge device and such. That's a pre-existing problem,
I believe, so not worrying about it.

The lookup from IP to box name is linear in the number of boxes,
because it currently uses the "compute service" to walk some JSON
files. That's not ideal, but it's probably harmless for now. Famous
last words.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: s0af8012705e98342k