exe: significant "ctr_host" cleanup
Because we went through docker and then nerdctl and now exelet, and we have
some complicated local setups, we've had a lot of overloaded behavior on things
called "ctr_host" which has caused me a bunch of strife and confusion.

Here is a description of the "new" world (as of this commit):

A given box is managed by an exelet. Exelet's are identified by host:port,
since they're a gRPC server. The "ctrhost" column of the boxes table is now
tcp://host:port to identify the relevant exelet. (I'd consider stripping out
the URL-essness of it, but it seems harmless enough.)

In the "dev" scenario, we let exed use -start-exelet to start an exelet,
and therefore we need to manage two hosts: the exed host (typically your
laptop) and the exelet VM. -start-exelet uses "lima-exe-ctr.local" as its
host. Due to the magic of mDNS, this is the name for the Lima VM, and SSH
works on it with port 22. setup_lima_hosts has configured your ~/.ssh/config
to use the key ~/.lima/_config/user.

In the "test" scenario, CTR_HOST is used as a pointer to a VM
where the test infra can start exelet. The format is ssh://user@host/
and the assumption is that passwordless SSH works. The github
workflows start a libvirt VM and configure this variable.
The default is ssh://lima-exe-ctr-tests/

SSH is used in two ways: (1) it's used to bootstrap exelet on
th machine that can run cloud-hypervisor and (2) it's used to
connect to those cloud-hypervisor VMs. In the second case,
we're using the same hostname, but the port is different, since
exelet has mapped the SSH port on the guest to a port on
the exelet host. We control the SSH settings in the second case
and the keys are in the database. We use this path for SSH
proper, for the HTTP proxy, and for the terminal UI.

Previously, there was some magic that diverged between our Lima configurations
and our CI configuration. Notably, Lima VMs magically map ports on the lima
VMs onto your Mac, which meant you could use "localhost" in a lot of places
where it didn't make sense. There was code to do that. We also mapped
a lot of names into IPs, and read SSH configuration to pick out ports and such.
SSH configuration can map any string (that is pretending to be a hostname but doesn't
have to resolve to anything) into any set of SSH settings, so you can do a lot of
magic, but the magic distracts.

Hopefully, by unifying CI and Mac dev, things become simpler. Famous last words.

If it turns out mdns doesn't work, we have some approaches: 1) we can come
up with environment variables for lima-exe-ctr and lima-exe-ctr-tests that
have IP addresses in them; 2) we can use /etc/hosts or similar; 3) we can
make DNS work somehow...