stage: add more Env bools to control behavior
Part of my crusade to eliminate devmode.

It's possible that not all of the bools picked here are correct.
It's mildly hard to tell because their values correlated strongly.
Also, a bunch of them are going to disappear shortly as we
add URLs to the Env struct.

Consider this a starting point for further refinement.