execore: filter and dedup wildcardDomains
Right now this obviously is unnecessary.
But we will shortly adjust the set of domains here
in order to support splitting box domains from web domains,
but also handle the case in which they are identical.
Doing this fairly clearly safe adjustment in a separate commit
here makes later commits smaller and easier to follow.