e1e: add restart and shutdown tests
As you can see from the golden tests and the TODOs in the code,
not everything is working perfectly.

But it works well enough to at least get the tests in,
so that we are aware of any changes in behavior here.
And so that we don't regress to anything worse,
such as hanging indefinitely after a restart,
or having a restarted box not come back up in a reachable state.

Fixes #70