add github action for building exelet
This adds a new GitHub action to build the exelet on a push to main
when the exelet is updated (e.g. exelet/*, cmd/exelet/*). This provides us
a standardized build of the exelet binary that includes arch specific
code (kernel, exe-init, sshd, sshd-session, sh, etc) that is embedded into
 the exelet binary.