exelet: use boxid-boxname as ids; run into SUN_LEN limit.
So, my VMs will now look like "vm6-azure-snow". The vm prefix is
arbitrary. the 6 is the ID in the databses in the boxes table. Names
themselves aren't unique since I think they're reusable after deletion,
so we shouldn't use them by themselves, but having them in there makes
spelunking easier.

This was all a bit tricky because when I finally uncovered the Cloud
Hypervisor error, it was:

 Fatal error: StartVmmThread(CreateApiServerSocket(Error { kind: InvalidInput, message: \"path must be shorter than SUN_LEN\" }

SUN_LEN is roughly 107.  This inclues the full path. So, it ends up
including some prefix, the vm name, some slashes, and
"cloud-hypervisor.sock". I renamed the latter to "chh.sock".
Our box names go up to 63 characters (boxname.Valid()). This was
previously 64, but DNS is up to 63 per RFC 1035

	2.3.4. Size limits

	Various objects and parameters in the DNS have size limits.  They are
	listed below.  Some could be easily changed, others are more
	fundamental.

	labels          63 octets or less

https://github.com/boldsoftware/exe/issues/105