all: finish box shard implementation
This commit finishes threading through the IP-based SNI for ssh,
so that you can do `ssh boxname.exe.xyz`, in addition to `ssh boxname@exe.xyz`.
We map the IP address back to the shard+user combination to infer boxname.

Some of the groundwork for this was previously laid, including
assigning shards to boxes and dealing with the EC2 1:1 NAT.

This commit adds local dev support and hooks it up to the piper plugin.

Local dev support works by adding a local mini route53 server, called alley53.
It sets up local DNS for exe.cloud subdomains, hooks 127.21.0.0/24 up to lo0,
and provides a simple HTML GUI interface for manual debugging.

The e1e tests will opportunistically use alley53 if it is running:

$ sudo go run ./cmd/alley53  # sorry about the sudo, unavoidable
$ go test ./e1e -run=Vanilla