zfs enhancements
Previously we were performing a resize to shrink the base image zvol
to account for compressed content and not all OCI images have uncompressed
size information. This was problematic because calculating the base
zvol snapshot was not accurate as the OCI image layer size doesn't account
for compressed content inside the image. In addition, we were exposed
to a DoS attack where users could request a huge image (e.g. 1T) and we
would pull to try to create the image. This changes the zfs base snapshots
in the following ways:

Base Snapshot Size
We no longer shrink the base image filesystems. Instead, we use a 10G sparse
base zvol as the base snapshot. We also check the image size early and fail
if it's larger than 10G. The sparse disk will only use what it needs and
we don't have to perform risky shrink operations.

Instance Disks
Instance disks behave like before and do copy-on-write as sparse zvols.

Load Operations
Previously we used the final "tank/sha256..." volume name during import. This
changes to use a temporary zvol as the import and then if successful, do an atomic
rename to the final image. This prevents corrupt or failed imported images from
being used or polluting the base image space.