all: use stage.Env.BoxDest more
Prompt:

stage.Env.BoxLogin was just added, in 625a9e2.
According to man ssh, the terminology for this is destination.
Please rename the method BoxDest, and use it in a more thoroughgoing way.
I know there are other places in the codebase where we construct this
either directly or as part of a larger string.