completeinit: new analyzer
Sometimes, as in recent commits, I want to extract parameter
sets out into a config struct, usually because they are long,
contain runs of identically typed fields, or contain bool params.

I am always nervous about doing that, however, because you lose
compiler validation that all call sites have been updated when you
add/alter a parameter.

The completeinit analyzer--which is 100% vibe-coded, I'm sorry--
helps alleviate these concerns.

It enforces that any struct annotated with //exe:completeinit
must be initialized with a struct literal with a complete set of keys.
It's not quite as good as compiler feedback, because it happens later,
but it does free us to use config structs with more confidence.

Use it on a couple of config structs that exist today.

Hat tip to Philip for vibe-coding the slogcontext analyzer,
which planted the idea that this was do-able.