e1e: in TestLoginWithExeFlow look in both hosts for cookies
The test used to look only for cookies on localhost.
That worked with Go 1.25 and earlier.
However, Go 1.26 changed cookies to be stored on the host
of the request (https://go.dev/issue/38988).
For TestLoginWithExeFlow that meant that one cookie was stored
on localhost and one was stored on what the test calls f.returnHost.
This change looks on both hosts for cookies, and works with both
Go 1.25 and 1.26.