execore: fix up JSON API
We still had box / box_name in the keys.

The shift to shards changed the correct shape of the API;
use ssh_dest instead of ssh_user+ssh_server.

This is a breaking change, but we're going to do it anyway,
because the old stuff is bad, and we don't have that many users yet.
The window is closing fast.