all: enable local manual multi-exelet testing
And introduce a server meta for setting a preferred exelet for new VMs.
This is a stopgap measure until we have proper resource-aware allocation.
Introduce /debug/exelet for controlling the preferred exelet.

To use, run `make run-devlets` -- note plural!
Only for manual testing.
Will definitely mess with your test runs;
prefer `make run-devlet` for normal use.

But this is enough to unblock experimentation
with multi-exelet setups.