passkeys: Suppress 'Resident credentials' passkey error from user UI
phil: so, this showed up in the e2e test report wherein I saw
this error in the screenshots. I don't know how many people use CHromium
without passkeys support, but Sketch does, so ... here we go.

Prompt: in a new worktree, if the passkey support for looking for
credentials isn't available, don't show that error the the user.
Currently it says "Resident credentials or empty 'allowCredentials'
lists are not supported at this time.", but we can console log that and
just not show the user our passkey support.

Some browsers report conditional UI support via isConditionalMediationAvailable()
but don't actually support resident credentials, causing the error:
'Resident credentials or empty allowCredentials lists are not supported at this time.'

Instead of showing this confusing error to users, we now just log it to the console
and gracefully degrade - the user can still use email authentication.