shelley: Reduce SSE bandwidth: send only new messages instead of all messages
phil: we used to be accidentally quadratic in the sense that we sent
all historical messages every time there was an event. This sucked.
(We also send too much data, but I wasn't convinced of that fix.)
So, now... let's not. I was going to tell the agent to just send
the id of the message and have the client handle it, but, honestly,
this seems to work...

phil: there was some weirdness with the usage data. It's too clever:
a 0 doesn't get serialized to JSON and the UI uses the previous value.
It's... fine.

Prompt: In a new worktree, investigate how much data we're sending to
the client during a conversation. Use the test harness code. I want to
make sure that in the results course of events every message gets sent
only once. And that only the display part of messages is sent. Evaluate
the situation and fix.

Send only new messages in SSE updates instead of re-sending all messages:
- Added notifySubscribersNewMessage() that publishes just the newly created message
- Changed recordMessage() to use the new function
- notifySubscribers() now only sends conversation metadata updates (for slug changes)
- Result: ~18x reduction in bandwidth (from 33KB to 1.8KB in test scenario)

Added test:
- TestMessageSentOnlyOnce: verifies each message is sent exactly once

TODO: Consider omitting llm_data when display_data is present for further bandwidth savings.